Intervals refactoring: avoid duplication code: IntervalsRecordingFragment is a version of IntervalsFragment.
