+++ src/main/java/de/dennisguse/opentracks/viewmodels/IntervalStatisticsModel.java
@@ -24,7 +24,6 @@ import de.dennisguse.opentracks.util.UnitConversions;
 public class IntervalStatisticsModel extends AndroidViewModel {
 
     private MutableLiveData<IntervalStatistics> intervalStats = new MutableLiveData<>();
-    private IntervalOption interval = IntervalOption.OPTION_1;
 
     public IntervalStatisticsModel(@NonNull Application application) {
         super(application);
